Moved Thought.jsx from Home page to components page, so it can be more easily accessed throughout the application.
