+++ src/components/Thought.jsx
